Move hashing.cpp to internal/c/parts/video/font to make gcc happy
